+++ core/data/src/main/java/com/wap/wapp/core/data/repository/survey/SurveyFormRepository.kt
@@ -1,0 +1,11 @@
+package com.wap.wapp.core.data.repository.survey
+
+import com.wap.wapp.core.model.survey.SurveyForm
+
+interface SurveyFormRepository {
+    suspend fun getSurveyForm(eventId: Int): Result<SurveyForm>
+
+    suspend fun getSurveyFormList(): Result<List<SurveyForm>>
+
+    suspend fun postSurveyForm(surveyForm: SurveyForm): Result<Unit>
+}
