+++ core/data/src/main/java/com/wap/wapp/core/data/repository/survey/SurveyRepository.kt
@@ -7,8 +7,4 @@ interface SurveyRepository {
     suspend fun getSurveyList(): Result<List<Survey>>
 
     suspend fun getSurvey(surveyId: String): Result<Survey>
-
-    suspend fun getSurveyFormList(): Result<List<SurveyForm>>
-
-    suspend fun getSurveyForm(eventId: Int): Result<SurveyForm>
 }
