+++ core/data/src/main/java/com/wap/wapp/core/data/repository/survey/SurveyRepositoryImpl.kt
@@ -1,7 +1,6 @@
 package com.wap.wapp.core.data.repository.survey
 
 import com.wap.wapp.core.model.survey.Survey
-import com.wap.wapp.core.model.survey.SurveyForm
 import com.wap.wapp.core.network.source.survey.SurveyDataSource
 import com.wap.wapp.core.network.source.user.UserDataSource
 import javax.inject.Inject
@@ -46,18 +45,4 @@ class SurveyRepositoryImpl @Inject constructor(
 
     // TODO 도메인 모델 구현을 위한 익스텐션, notice DataSource 구현 후 소거
     private fun String.toDomain(): String = this
-
-    override suspend fun getSurveyFormList(): Result<List<SurveyForm>> {
-        return surveyDataSource.getSurveyFormList().mapCatching { surveyFormResponseList ->
-            surveyFormResponseList.map { surveyFormResponse ->
-                surveyFormResponse.toDomain()
-            }
-        }
-    }
-
-    override suspend fun getSurveyForm(eventId: Int): Result<SurveyForm> {
-        return surveyDataSource.getSurveyForm(eventId).mapCatching { surveyFormResponse ->
-            surveyFormResponse.toDomain()
-        }
-    }
 }
